[#51] Long으로 변환

+++ src/main/java/com/depromeet/breadmapbackend/bakeries/controller/BakeriesController.java
@@ -32,7 +32,7 @@ public class BakeriesController {
      */
     @ApiOperation(value = "단일 빵집 리뷰 리스트", notes = "단일 빵집에 있는 메뉴에 대한 리뷰 리스트 조회")
     @GetMapping(value = "/{bakeryId}/review")
-    public ResponseEntity<ReviewsListResponse> getBakeryReviewList(@PathVariable Integer bakeryId){
+    public ResponseEntity<ReviewsListResponse> getBakeryReviewList(@PathVariable Long bakeryId){
         return null;
     }
 
@@ -42,7 +42,7 @@ public class BakeriesController {
      */
     @ApiOperation(value = "단일 빵집 메뉴 리스트", notes = "단일 빵집에 있는 메뉴 리스트 조회")
     @GetMapping(value = "/{bakeryId}/menus")
-    public ResponseEntity<BakeryMenusListResponse> getBakeryMenusList(@PathVariable Integer bakeryId){
+    public ResponseEntity<BakeryMenusListResponse> getBakeryMenusList(@PathVariable Long bakeryId){
         return null;
     }
 
@@ -52,7 +52,7 @@ public class BakeriesController {
      */
     @ApiOperation(value = "빵집 별점 넣기", notes = "빵집 별점 넣기")
     @PostMapping(value = "/{bakeryId}/rating")
-    public ResponseEntity postBakeryRating(@PathVariable Integer bakeryId, @RequestBody PostBakeryRatingRequest postBakeryRatingRequest){
+    public ResponseEntity postBakeryRating(@PathVariable Long bakeryId, @RequestBody PostBakeryRatingRequest postBakeryRatingRequest){
         return ApiResponse.created(null);
     }
